Changed quantities in discount and calculator to long.

+++ src/main/java/com/wrist/source2sea/domain/Discount.java
@@ -9,10 +9,10 @@ public class Discount {
     private static final String INVALID_QUANTITY_MESSAGE = "Invalid discount quantity: %d. It must be less than 2 !";
     private static final String INVALID_PRICE_FOR_QUANTITY = "Invalid discount price for quantity: %.2f. It must be greater than 0 !";
 
-    Integer quantity;
+    Long quantity;
     BigDecimal priceForQuantity;
 
-    public Discount(Integer quantity, BigDecimal priceForQuantity) {
+    public Discount(Long quantity, BigDecimal priceForQuantity) {
         assertDiscountQuantity(quantity);
         assertDiscountPriceForQuantity(priceForQuantity);
 
@@ -20,7 +20,7 @@ public class Discount {
         this.priceForQuantity = priceForQuantity;
     }
 
-    private void assertDiscountQuantity(Integer quantity) {
+    private void assertDiscountQuantity(Long quantity) {
         if (quantity < 2) {
             throw new IllegalArgumentException(String.format(INVALID_QUANTITY_MESSAGE, quantity));
         }
